pc-thermometer-desktop-service/requirements.txt
altgraph==0.17.4
cffi==1.17.1
clr_loader==0.2.7.post0
packaging==24.2
pefile==2023.2.7
pycparser==2.22
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2025.1
pyserial==3.5
pythonnet==3.0.5
pywin32-ctypes==0.2.3
setuptools==75.8.0
